% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mesh_thesaurus.R
\name{data_mesh_thesuarus}
\alias{data_mesh_thesuarus}
\title{Download and Combine MeSH and Supplemental Thesauruses}
\usage{
data_mesh_thesuarus(force_download = FALSE)
}
\arguments{
\item{force_download}{A logical value indicating whether to force re-downloading 
of the data even if it already exists locally.}
}
\value{
A data.table containing the combined MeSH and supplemental thesaurus data.
}
\description{
This function downloads and combines the MeSH (Medical Subject Headings) Thesaurus 
and a supplemental concept thesaurus for use in biomedical research and analysis.
The data is sourced from specified URLs and stored locally for subsequent use.
}
\examples{
\donttest{
if (interactive()) {
  # Code that downloads data or performs other interactive-only operations
  data <- data_mesh_thesaurus()
}
}
}
