% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pmc_list.R
\name{data_pmc_list}
\alias{data_pmc_list}
\title{Download and Process PMC Open Access File List}
\usage{
data_pmc_list(force_install = FALSE)
}
\arguments{
\item{force_install}{Logical, if TRUE, forces the re-download and processing of 
the file even if it already exists locally. Default is FALSE.}
}
\value{
A data frame containing the processed PMC open access file list.
}
\description{
This function downloads the PubMed Central (PMC) open access file list from the
National Center for Biotechnology Information (NCBI) and processes it for use.
The list is saved locally. If the file does not exist or if `force_install` is TRUE, 
it will be downloaded and processed.
}
\examples{
\donttest{
if (interactive()) {
  # Code that downloads data or performs other interactive-only operations
  data <- data_pmc_list()
}
}
}
