% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_pubmed.R
\name{.extract_basic}
\alias{.extract_basic}
\title{Extract Basic Information from PubMed Records}
\usage{
.extract_basic(g)
}
\arguments{
\item{g}{An XML node set representing a single PubMed record.}
}
\value{
A named vector with basic bibliographic information from a PubMed record.
}
\description{
An internal function that parses XML records from PubMed. It extracts essential bibliographic information such as PubMed ID, journal title, article title, publication year, and abstract.
}
