% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mesh_embeddings.R
\name{data_mesh_embeddings}
\alias{data_mesh_embeddings}
\title{Download and Process Mesh and SCR Embeddings}
\usage{
data_mesh_embeddings()
}
\value{
A data frame containing the processed Mesh and SCR embeddings data.
}
\description{
This function downloads Mesh and SCR embeddings data from the specified URLs and processes it for use.
The data is saved locally in RDS format. If the files do not exist, they will be downloaded and processed.
}
\examples{
\donttest{
if (interactive()) {
  # Code that downloads data or performs other interactive-only operations
  data <- data_mesh_embeddings()
}
}

}
